Remove Bid transactions since they are not used
Fix a bug when a transaction writes a key it previously read.  It's
possible the key was written by a different transaction between the
read and the lock, so preserve the key version when locking so I can
read-validate.

This only surfaced during the non-commutative transactions which I
just started running yesterday.